refactor to allow the web app to roll dice.
