--- roll_witch/dice_bot/operation/__init__.py
@@ -1,5 +1,0 @@
-from roll_witch.dice_bot.operation.operations.standard import RollOperation
-from .parser import OperationParser
-from .result import OperationResult
-
-__all__ = ["RollOperation", "OperationParser", "OperationResult"]
