+++ roll_witch/rolling/operation/__init__.py
@@ -1,0 +1,4 @@
+from .parser import OperationParser
+from .result import OperationResult
+
+__all__ = ["OperationParser", "OperationResult"]
