+++ roll_witch/rolling/operation/input/parser/base.py
