+++ roll_witch/rolling/operation/input/simple_regex.py
