+++ roll_witch/rolling/operation/input/spec/basic.py
@@ -2,8 +2,8 @@ import re
 from math import ceil
 from typing import Match
 
-from roll_witch.dice_bot.operation.input.spec import InputPartSpec
-from roll_witch.dice_bot.roller import RollSpec
+from .base import InputPartSpec
+from roll_witch.rolling.roller import RollSpec
 
 
 class BasicRpgSpec(InputPartSpec):
