+++ roll_witch/rolling/operation/output/base.py
@@ -1,5 +1,5 @@
-from roll_witch.dice_bot.roller import RollResult
-from roll_witch.dice_bot.protocols import Result
+from roll_witch.rolling.roller import RollResult
+from roll_witch.rolling.protocols import Result
 from abc import ABC, abstractmethod
 from typing import Protocol
