+++ roll_witch/rolling/protocols/operation_factory.py
