+++ roll_witch/rolling/protocols/result.py
