+++ tests/test_output_parser.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
-from roll_witch.dice_bot.operation.output import StandardOutputWriter, TargetedOutputWriter
-from roll_witch.dice_bot.operation.spec import RollSpec
-from roll_witch.dice_bot.roller import RollResult
+from roll_witch.rolling.operation.output import StandardOutputWriter, TargetedOutputWriter
+from roll_witch.rolling.operation.spec import RollSpec
+from roll_witch.rolling.roller import RollResult
 
 
 class TestStandardOutputWriter(TestCase):
