+++ tests/test_roll_result.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
-from roll_witch.dice_bot.operation.spec import RollSpec
-from roll_witch.dice_bot.roller import RollResult
+from roll_witch.rolling.operation.spec import RollSpec
+from roll_witch.rolling.roller import RollResult
 
 
 class TestRollResult(TestCase):
