+++ tests/test_token_output_writer.py
@@ -1,8 +1,8 @@
 from unittest import TestCase
-from roll_witch.dice_bot.operation.output import OperationOutputWriter
-from roll_witch.dice_bot.operation.spec import OperationSpec, RollSpec
-from roll_witch.dice_bot.roller import RollResult
-from  roll_witch.dice_bot.operation import OperationResult
+from roll_witch.rolling.operation.output import OperationOutputWriter
+from roll_witch.rolling.operation.spec import OperationSpec, RollSpec
+from roll_witch.rolling.roller import RollResult
+from roll_witch.rolling.operation import OperationResult
 
 
 class TestTokenOutputWriter(TestCase):
